% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cg_api_url}
\alias{cg_api_url}
\title{Helper function for base CoinCap API URL}
\usage{
cg_api_url(endpoint)
}
\arguments{
\item{endpoint}{character}
}
\value{
character API URL
}
\description{
Helper function for base CoinCap API URL
}
