% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{cg_coin_markets_summ}
\alias{cg_coin_markets_summ}
\title{Query for market summary data}
\usage{
cg_coin_markets_summ(
  coin_ct = 250,
  vs_currency = "usd",
  category,
  order = "market_cap_desc",
  price_change_pct = c("1h", "24h", "7d", "14d", "30d", "200d", "1y")
)
}
\arguments{
\item{coin_ct}{the number of rows to return}

\item{vs_currency}{quote currency}

\item{category}{coin category to query for}

\item{order}{the desired row order of the return data frame}

\item{price_change_pct}{time ranges to include returns for}
}
\value{
\code{data.frame}
}
\description{
Query for market summary data
}
