% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{cg_coin_snapshot}
\alias{cg_coin_snapshot}
\title{Query for OHLC coin data}
\usage{
cg_coin_snapshot(coin = "bitcoin", date = format(Sys.Date(), "\%d-\%m-\%Y"))
}
\arguments{
\item{coin}{coin to query for}

\item{date}{date of snapshot to query for}
}
\value{
data frame
}
\description{
Query for OHLC coin data
}
