% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{cg_market_chart_range}
\alias{cg_market_chart_range}
\title{Query for high-level market data by time range}
\usage{
cg_market_chart_range(
  coin = "bitcoin",
  vs_currency = "usd",
  start_tm = Sys.time() - lubridate::days(30),
  end_tm = Sys.time()
)
}
\arguments{
\item{coin}{coin to query for}

\item{vs_currency}{quote currency}

\item{start_tm}{min time for desired time range}

\item{end_tm}{max time for desired time range}
}
\value{
\code{data.frame}
}
\description{
Query for high-level market data by time range
}
