% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{cg_ohlc}
\alias{cg_ohlc}
\title{Query for OHLC coin data}
\usage{
cg_ohlc(coin = "bitcoin", vs_currency = "usd", day_ct = 30)
}
\arguments{
\item{coin}{coin to query for}

\item{vs_currency}{quote currency}

\item{day_ct}{count of days of history to query for}
}
\value{
\code{data.frame}
}
\description{
Query for OHLC coin data
}
\details{
Candle's body:
\itemize{
\item 1 - 2 days: 30 minutes
\item 3 - 30 days: 4 hours
\item 31 and before: 4 days
}
}
