% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{cg_supported_vs_currencies}
\alias{cg_supported_vs_currencies}
\title{Query for supported vs currencies}
\usage{
cg_supported_vs_currencies()
}
\value{
vector
}
\description{
Query for supported vs currencies
}
