% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{datetime_millisec}
\alias{datetime_millisec}
\alias{millisec_to_datetime}
\alias{datetime_to_millisec}
\title{Helper functions for converting between millisecond numerics and datetime objects}
\usage{
millisec_to_datetime(x)

datetime_to_millisec(x)
}
\arguments{
\item{x}{numeric/datetime}
}
\value{
numeric/datetime
}
\description{
Helper functions for converting between millisecond numerics and datetime objects
}
