% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gg_crypto_platform_prop}
\alias{gg_crypto_platform_prop}
\title{Query for stablecoin data and plot pie chart of market dominance}
\usage{
gg_crypto_platform_prop()
}
\value{
ggplot
}
\description{
Query for stablecoin data and plot pie chart of market dominance
}
