% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{gg_stablecoins_pie}
\alias{gg_stablecoins_pie}
\title{Query for stablecoin data and plot pie chart of market dominance}
\usage{
gg_stablecoins_pie()
}
\value{
\code{data.frame}
}
\description{
Query for stablecoin data and plot pie chart of market dominance
}
